feat: add playRound method to GameController
Filled out switchPlayerTurn method.

Adjusted placeMarker method on Gameboard to check for valid placement,
and return boolean to GameController when called.

Added getMarker method to Player to be called when a player is taking
their turn.

Simplified simple 'get' methods to arrow functions to utilize implicit
return. This thins out the code and improves readability.