fix: add value arg to 1,1 diag forEach()
When incrementing the markersInDiag array for a marker at middle-middle,
the forEach uses the 'index' arg. But it was missing the 'value' arg,
which caused forEach to use 'index' to represent 'value'.

Also fixed drawStatus key notation. Since winCondition was moved outside
of the checkGameOver method, drawStatus became a defined property.